+++ axelor-base/src/main/java/com/axelor/apps/base/web/UserController.java
@@ -177,4 +177,13 @@ public class UserController {
       TraceBackService.trace(response, e);
     }
   }
+
+  public void setDefaultUserMethod(ActionRequest request, ActionResponse response) {
+    try {
+      AppSettings appSettings = AppSettings.get();
+      response.setValue("language", appSettings.get("application.locale"));
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
+  }
 }