+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add global tracking log feature.
 
 ## Improvements
+- USER : Default User language is based on application.locale from application.properties
 - BASE : Cache memory performace improved by not stocking geographical entities anymore
 
 ## Bug Fixes